requirements.txt
django>=1.8
django-cms>=3.0
django-filer>=1.1.0
django-appconf>=1.0.1
